start an optional implementation for async_trait
we add a feature flag (which is off by default) to gate the whether you
want an async trait or not
