+++ Cargo.toml
@@ -10,9 +10,14 @@ categories = ["web-programming"]
 license = "GPL-3.0"
 edition = "2018"
 
+[features]
+default = []
+async = ["async-trait"]
+
 [dependencies]
 reqwest = { version = "0.10", features = [ "json" ] }
 serde = { version = "1.0", features = [ "derive" ] }
+async-trait = {version = "*", optional = true}
 
 [dev-dependencies]
 serde_json = "1.0"
