+++ src/lib.rs
@@ -3,7 +3,17 @@
 //! Use [`resolve`] to fetch remote resources, and [`Resolver`] to serve your own resources.
 
 use reqwest::{header::ACCEPT, Client};
-use serde::{Serialize, Deserialize};
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "async")]
+mod async_trait;
+#[cfg(feature = "async")]
+pub use crate::async_trait::*;
+
+#[cfg(not(feature = "async"))]
+mod sync_trait;
+#[cfg(not(feature = "async"))]
+pub use crate::sync_trait::*;
 
 #[cfg(test)]
 mod tests;
@@ -139,7 +149,10 @@ pub async fn resolve_with_prefix(
 /// Fetches a Webfinger resource.
 ///
 /// If the resource doesn't have a prefix, `acct:` will be used.
-pub async fn resolve(acct: impl Into<String>, with_https: bool) -> Result<Webfinger, WebfingerError> {
+pub async fn resolve(
+    acct: impl Into<String>,
+    with_https: bool,
+) -> Result<Webfinger, WebfingerError> {
     let acct = acct.into();
     let mut parsed = acct.splitn(2, ':');
     let first = parsed.next().ok_or(WebfingerError::ParseError)?;
@@ -168,38 +181,4 @@ pub enum ResolverError {
 
     /// The requested resource was not found.
     NotFound,
-}
-
-/// A trait to easily generate a WebFinger endpoint for any resource repository.
-///
-/// The `R` type is your resource repository (a database for instance) that will be passed to the
-/// [`find`](Resolver::find) and [`endpoint`](Resolver::endpoint) functions.
-pub trait Resolver<R> {
-    /// Returns the domain name of the current instance.
-    fn instance_domain<'a>(&self) -> &'a str;
-
-    /// Tries to find a resource, `acct`, in the repository `resource_repo`.
-    ///
-    /// `acct` is not a complete `acct:` URI, it only contains the identifier of the requested resource
-    /// (e.g. `test` for `acct:test@example.org`)
-    ///
-    /// If the resource couldn't be found, you may probably want to return a [`ResolverError::NotFound`].
-    fn find(&self, prefix: Prefix, acct: String, resource_repo: R) -> Result<Webfinger, ResolverError>;
-
-    /// Returns a WebFinger result for a requested resource.
-    fn endpoint(&self, resource: impl Into<String>, resource_repo: R) -> Result<Webfinger, ResolverError> {
-        let resource = resource.into();
-        let mut parsed_query = resource.splitn(2, ':');
-        let res_prefix = Prefix::from(parsed_query.next().ok_or(ResolverError::InvalidResource)?);
-        let res = parsed_query.next().ok_or(ResolverError::InvalidResource)?;
-
-        let mut parsed_res = res.splitn(2, '@');
-        let user = parsed_res.next().ok_or(ResolverError::InvalidResource)?;
-        let domain = parsed_res.next().ok_or(ResolverError::InvalidResource)?;
-        if domain == self.instance_domain() {
-            self.find(res_prefix, user.to_string(), resource_repo)
-        } else {
-            Err(ResolverError::WrongDomain)
-        }
-    }
 }
